requirements.txt
python-dotenv
python-telegram-bot==13.15
TikTokApi
python-amazon-paapi
